aws/genai/aws_sagemaker/langgraph_agent_deploy/requirements.txt
langgraph==1.0.7
boto3==1.42.39
fastapi==0.128.0
uvicorn==0.34.0
httpx==0.28.1
pydantic==2.10.6
python-dotenv==1.0.0
openai==1.58.1
sagemaker
